requirements.txt
fastapi
PyMySQL
SQLAlchemy
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
